+++ src/Oro/Bundle/HelpBundle/Tests/Unit/Fixtures/Bundles/BarBundle/BarBundle.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Oro\Bundle\HelpBundle\Tests\Unit\DependencyInjection\Fixtures\BarBundle;
+namespace Oro\Bundle\HelpBundle\Tests\Unit\Fixtures\Bundles\BarBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;