+++ CHANGELOG.md
@@ -77,6 +77,9 @@ The current file describes significant changes in the code that may affect the u
 #### FilterBundle
 * The `datasource` attribute for `oro_filter.extension.orm_filter.filter` tag has been removed as it is redundant.
 
+#### HelpBundle
+* The possibility to define `resources`, `vendors` and `routes` sections via HelpBundle configuration has been removed. Use `Resources/config/oro/help.yml` instead.
+
 #### MessageQueueBundle
 * The `DefaultTransportFactory` and related configuration option `oro_message_queue.transport.defaut` was removed. Check `config/config.yml` in your application.